Changes to STDOUT and STDERR text stream processing.
Static method Executable.process_stream() now opens a file handle in text mode, if a file path was provided.
If no file path was provided, it will write to STDOUT or STDERR depending on the file type option, so that from now on, STDERR and STDOUT will be written to separate streams.
Introduced the private function get_timestamp() to simplify code.
The timestamp is now only prepended if a debug level was set to allow for cleaner downstream parsing.