Further re-engineering towards concurrent processes.
Moved the majority of code from the bsf.runnables.generic.run() function into a new Runnable.run_consecutively() method, so that it can be re-used for the ConsecutiveRunnable class, as well as for pre-run and post-run lists of the ConcurrentRunnable class.
In class ConcurrentRunnable, added support for lists of RunnableStep objects run before or after the concurrent processes. This allows simple tasks like creating directories, named pipes or compressing output files.
In module bsf.runnables.concurrent, check child process return codes and raise Exceptions if required.