Adjustments to changes in the Aligner class.
In class FilePathSummary changed the names of instance variables for summary tables, to better indicate the TSV format.
In method StarAligner.add_runnable_step_summary(), changed the name of the RunnableStep to avoid conflicts with the RunnableStep summarising Picard Alignment Summary Metrics reports in the Aligner class.
Fixed type hinting in Python docstrings.