Automatise the DESeq analysts.
Submit processes for all designs of the rnaseq_deseq_analysis stage.
If a contrast annotation sheet was provided and contains records for a particular design also submit a process for the rnaseq_deseq_results stage.
No longer raise an Exception, if a contrast annotation sheet was not provided.
In case the transcriptome_gtf_path instance variable is not defined, default to StandardFilePath.get_resource_transcriptome_gtf().