Add pre-exec infrastructure.
In class Stage add an instance variable and corresponding configuration option to define a template script.
In method Analysis.set_stage_runnable(), insert an Executable object for the template script if one is defined.
In class StandardFilePath add method get_template_scripts() that provides the value of the   template_scripts option of the directories section of the global configuration file.