Updated the copyright notice.
Deleted the LGPL3 text from the PyCharm copyright template, to allow updates of just the singe line copyright notice.
Rearranged the file header to move the copyright notice and the LGPL3 licence to the top of the file header so that PyCharm can update.
Removed the Python source encoding line from the remaining GNU Bash scripts.